refactor(tools): make the delegation/ split self-contained (post-#6648)
Address #6650 (delegation findings):
1. Move WorkflowAgentInputSchema and its `WorkflowAgentInput` type from
   DelegationTools.ts into delegation/inputFields.ts (beside memoriesField,
   which the schema uses), removing the child→parent `import type` back-ref
   so inputFields.ts is self-contained. DelegationTools.ts imports the schema
   and re-exports the type, so its public surface is unchanged.
2. Rename subagentExecution.ts's LOG_CHANNEL from the stale parent name
   'DelegationTools' to the module-scoped 'delegation'.

Pure move + a constant rename; typecheck clean, DelegationTools tests pass,
public export set unchanged. (Finding 3 — codex schemas imported into the
webview frontend — is a separate codexShared concern, left for its own change.)
