+++ src/tools/delegation/inputFields.ts
@@ -22,9 +22,6 @@ import { AbsoluteFS, WorkspaceFS } from '@utils/files';
 import { hasExtension } from '@utils/core/pathCore';
 import { isNonEmptyString } from '@utils/core/stringCore';
 
-// Local imports - delegation
-import type { WorkflowAgentInput } from '../DelegationTools';
-
 const LARGE_BIB_LIMIT_BYTES = 100 * 1024;
 
 /**
@@ -53,6 +50,59 @@ export const memoriesField = z
       }
     }
   });
+
+/** Schema for the delegate_workflow tool (document processing). */
+export const WorkflowAgentInputSchema = z.strictObject({
+  agent: z.string().describe('Name of the workflow agent to execute'),
+  model: z
+    .string()
+    .nullish()
+    .describe(
+      'Model short name from the Available models line. Omit unless the user explicitly requested a model; defaults to the current model when available.',
+    ),
+  instruction: z
+    .string()
+    .describe(
+      'What the agent should do, in plain prose. If you attach context or media files, name each one and say what role it plays — e.g., "preamble.tex defines the math macros; refs.bib is the bibliography to cite from; figure.png shows the panel layout to match". The sub-agent has no other signal for why each file was attached.',
+    ),
+  inputFiles: z
+    .array(z.string())
+    .min(1)
+    .describe(
+      'Files the agent rewrites. List every file you want it to touch. The agent emits one revised <document> per entry.',
+    ),
+  contextFiles: z
+    .array(z.string())
+    .prefault([])
+    .describe(
+      'Read-only context the agent should see but not modify: guidance, examples, related papers, bibliographies (.bib), style/macro definitions (.sty/.cls). Explain each one in the instruction.',
+    ),
+  mediaFiles: z
+    .array(z.string())
+    .prefault([])
+    .describe('Images, figures, PDFs, or audio files the agent should view.'),
+  extractFigures: z
+    .boolean()
+    .nullish()
+    .describe(
+      'When true, automatically extracts figures referenced by the input LaTeX file(s) (via \\includegraphics, \\begin{overpic}) and attaches them as media files. Merges with any explicitly provided mediaFile/mediaFiles.',
+    ),
+  extractTikz: z
+    .boolean()
+    .nullish()
+    .describe(
+      'When true, extracts TikZ figures from the input LaTeX file(s), compiles them into standalone PDFs, and attaches them as media files.',
+    ),
+  outputFiles: z
+    .array(z.string())
+    .prefault([])
+    .describe(
+      'Output file paths. Must be a subset of input files—never create new files or change format. Leave empty for default suffix-based outputs.',
+    ),
+  memories: memoriesField,
+});
+
+export type WorkflowAgentInput = z.infer<typeof WorkflowAgentInputSchema>;
 
 const WORKTREE_DISABLED_MESSAGE =
   "git worktree support is disabled in this workspace. Omit working_directory, or enable 'Allow agents to work in git worktrees' on the Multi-Agent settings tab.";
